ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.4
langchain-groq==0.3.4
langchain-core==0.3.68
python-dotenv==1.1.1
sse-starlette==2.3.6
